+++ piecrust/src/session.rs
@@ -57,6 +57,7 @@ unsafe impl Sync for Session {}
 /// [`transacted`]: Session::transact
 /// [`commit`]: Session::commit
 /// [`set_meta`]: Session::set_meta
+#[derive(Debug)]
 pub struct Session {
     call_stack: CallStack,
     instance_map: BTreeMap<ModuleId, (*mut WrappedInstance, u64)>,
@@ -112,31 +113,10 @@ impl Session {
     ///
     /// [`ModuleId`]: ModuleId
     /// [`deploy_with_id`]: `Session::deploy_with_id`
-    pub fn deploy(&mut self, bytecode: &[u8]) -> Result<ModuleId, Error> {
-        let hash = blake3::hash(bytecode);
-        let module_id = ModuleId::from_bytes(hash.into());
-
-        self.deploy_with_id(module_id, bytecode)?;
-
-        Ok(module_id)
-    }
-
-    /// Deploy a module, returning its [`ModuleId`]. The ID is computed using a
-    /// `blake3` hash of the `bytecode`.
-    /// If contract exports an initialization method, it will be called with
-    /// the arguments provided.
-    ///
-    /// If one needs to specify the ID, [`deploy_with_id`] is available.
-    /// If one would like to *not* call the initialization method, [`deploy`] is
-    /// available.
-    ///
-    /// [`ModuleId`]: ModuleId
-    /// [`deploy_with_id`]: `Session::deploy_with_id`
-    /// [`deploy`]: `Session::deploy`
-    pub fn deploy_and_init<Arg>(
+    pub fn deploy<Arg>(
         &mut self,
         bytecode: &[u8],
-        arg: &Arg,
+        arg: Option<Arg>,
     ) -> Result<ModuleId, Error>
     where
         Arg: for<'b> Serialize<StandardBufSerializer<'b>>,
@@ -144,7 +124,7 @@ impl Session {
         let hash = blake3::hash(bytecode);
         let module_id = ModuleId::from_bytes(hash.into());
 
-        self.deploy_with_id_and_init(module_id, bytecode, arg)?;
+        self.deploy_with_id(module_id, bytecode, arg)?;
 
         Ok(module_id)
     }
@@ -155,11 +135,28 @@ impl Session {
     /// available.
     ///
     /// [`deploy`]: `Session::deploy`
-    pub fn deploy_with_id(
+    pub fn deploy_with_id<Arg>(
         &mut self,
         id: ModuleId,
         bytecode: &[u8],
-    ) -> Result<(), Error> {
+        arg: Option<Arg>,
+    ) -> Result<(), Error>
+    where
+        Arg: for<'b> Serialize<StandardBufSerializer<'b>>,
+    {
+        self.do_deploy_with_id(id, bytecode, arg, None)
+    }
+
+    fn do_deploy_with_id<Arg>(
+        &mut self,
+        id: ModuleId,
+        bytecode: &[u8],
+        arg: Option<Arg>,
+        ser_arg: Option<Vec<u8>>,
+    ) -> Result<(), Error>
+    where
+        Arg: for<'b> Serialize<StandardBufSerializer<'b>>,
+    {
         if !self.module_session.module_deployed(id) {
             let wrapped_module =
                 WrappedModule::new(bytecode, None::<Objectcode>)?;
@@ -170,9 +167,27 @@ impl Session {
 
         self.create_instance(id)?;
 
+        let instance = self.instance(&id).expect("instance should exist");
+
+        if !matches!((&arg, &ser_arg), (None, None))
+            && !instance.is_function_exported(CONTRACT_INIT_METHOD)
+        {
+            return Err(InitalizationError(
+                "deploy initialization failed as init method is not exported"
+                    .into(),
+            ));
+        }
+
+        let s_arg = match (arg, &ser_arg) {
+            (Some(a), _) => Some(self.initialize::<Arg>(id, &a, None)?),
+            (None, Some(_)) => Some(self.initialize::<()>(id, &(), ser_arg)?),
+            _ => None,
+        };
+
         self.call_history.push(From::from(Deploy {
             module_id: id,
             bytecode: bytecode.to_vec(),
+            ser_arg: s_arg,
         }));
 
         Ok(())
@@ -243,7 +258,7 @@ impl Session {
         Ret::Archived: Deserialize<Ret, Infallible>
             + for<'b> CheckBytes<DefaultValidator<'b>>,
     {
-        if !self.is_deploy() && method_name == CONTRACT_INIT_METHOD {
+        if method_name == CONTRACT_INIT_METHOD {
             return Err(InitalizationError("init call not allowed".into()));
         }
 
@@ -298,7 +313,7 @@ impl Session {
         Ret::Archived: Deserialize<Ret, Infallible>
             + for<'b> CheckBytes<DefaultValidator<'b>>,
     {
-        if !self.is_deploy() && method_name == CONTRACT_INIT_METHOD {
+        if method_name == CONTRACT_INIT_METHOD {
             return Err(InitalizationError("init call not allowed".into()));
         }
 
@@ -322,6 +337,46 @@ impl Session {
         let ret = ta.deserialize(&mut Infallible).expect("Infallible");
 
         Ok(ret)
+    }
+
+    pub fn initialize<Arg>(
+        &mut self,
+        module: ModuleId,
+        arg: &Arg,
+        ser_arg: Option<Vec<u8>>,
+    ) -> Result<Vec<u8>, Error>
+    where
+        Arg: for<'b> Serialize<StandardBufSerializer<'b>>,
+    {
+        let mut call = Call {
+            ty: CallType::T,
+            module,
+            fname: CONTRACT_INIT_METHOD.to_string(),
+            fdata: vec![],
+            limit: self.limit,
+        };
+        let s_arg = match ser_arg {
+            None => {
+                let mut sbuf = [0u8; SCRATCH_BUF_BYTES];
+                let scratch = BufferScratch::new(&mut sbuf);
+                let ser = BufferSerializer::new(&mut self.buffer[..]);
+                let mut ser =
+                    CompositeSerializer::new(ser, scratch, Infallible);
+
+                ser.serialize_value(arg).expect("Infallible");
+                let pos = ser.pos();
+                call.fdata = self.buffer[..pos].to_vec();
+                call.fdata.to_vec()
+            }
+            Some(s_arg) => {
+                call.fdata = s_arg.to_vec();
+                s_arg
+            }
+        };
+
+        self.execute_until_ok(call)?;
+
+        Ok(s_arg)
     }
 
     /// Return the state root of the current state of the session.
@@ -678,7 +733,7 @@ impl Session {
                     res = self.call_if_not_error(call);
                 }
                 CallOrDeploy::Deploy(deploy) => {
-                    self.deploy_with_id(deploy.module_id, &deploy.bytecode)
+                    self.do_deploy_with_id::<()>(deploy.module_id, &deploy.bytecode, None, deploy.ser_arg)
                         .expect("Only deploys that succeed should be added to the history");
                 }
             }
@@ -759,6 +814,7 @@ impl From<Deploy> for CallOrDeploy {
 struct Deploy {
     module_id: ModuleId,
     bytecode: Vec<u8>,
+    ser_arg: Option<Vec<u8>>,
 }
 
 #[derive(Debug)]