+++ piecrust/src/store.rs
@@ -41,6 +41,7 @@ const OBJECTCODE_EXTENSION: &str = ".a";
 type Root = [u8; ROOT_LEN];
 
 /// A store for all module commits.
+#[derive(Debug)]
 pub struct ModuleStore {
     sync_loop: thread::JoinHandle<()>,
     call: mpsc::Sender<Call>,
@@ -277,7 +278,7 @@ fn index_from_path<P: AsRef<Path>>(
     Ok(modules)
 }
 
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub(crate) struct Commit {
     modules: BTreeMap<ModuleId, Root>,
     diffs: BTreeSet<ModuleId>,