+++ piecrust/tests/crossover.rs
@@ -26,8 +26,16 @@ fn crossover() -> Result<(), Error> {
     let mut session = vm.genesis_session();
     session.set_point_limit(u64::MAX / 100);
 
-    session.deploy_with_id(CROSSOVER_ONE, module_bytecode!("crossover"))?;
-    session.deploy_with_id(CROSSOVER_TWO, module_bytecode!("crossover"))?;
+    session.deploy_with_id(
+        CROSSOVER_ONE,
+        module_bytecode!("crossover"),
+        None::<()>,
+    )?;
+    session.deploy_with_id(
+        CROSSOVER_TWO,
+        module_bytecode!("crossover"),
+        None::<()>,
+    )?;
 
     // These value should not be set to `INITIAL_VALUE` in the contract.
     const CROSSOVER_TO_SET: i32 = 42;