+++ piecrust/tests/deploy_with_id.rs
@@ -15,7 +15,7 @@ pub fn deploy_with_id() -> Result<(), Error> {
     let some_id = [1u8; 32];
     let module_id = ModuleId::from(some_id);
     let mut session = vm.genesis_session();
-    session.deploy_with_id(module_id, bytecode)?;
+    session.deploy_with_id(module_id, bytecode, None::<()>)?;
 
     assert_eq!(
         session.query::<(), i64>(module_id, "read_value", &())?,