+++ piecrust/tests/vector.rs
@@ -12,7 +12,7 @@ pub fn vector_push_pop() -> Result<(), Error> {
 
     let mut session = vm.genesis_session();
 
-    let id = session.deploy(module_bytecode!("vector"))?;
+    let id = session.deploy(module_bytecode!("vector"), None::<()>)?;
 
     const N: usize = 128;